[yurong] refactor: change trainee id type to long and do some refactor

+++ src/main/java/com/example/demo/repository/TraineeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TraineeRepository extends CrudRepository<Trainee, Integer> {
+public interface TraineeRepository extends CrudRepository<Trainee, Long> {
 }
